move the rest of scalachess.Game import code over to lila.tree
