+++ modules/core/src/main/game/misc.scala
@@ -161,31 +161,11 @@ trait Namer:
   def gameVsText(game: Game, withRatings: Boolean = false)(using lightUser: LightUser.Getter): Fu[String]
   def playerText(player: Player, withRating: Boolean = false)(using lightUser: LightUser.Getter): Fu[String]
 
-case class TagResult(status: Status, winner: Option[Color]):
-  // duplicated from Game.finish
-  def finished = status >= Status.Mate
-
-case class ImportData(pgn: PgnStr, analyse: Option[String])
-case class ImportReady(game: NewGame, replay: Replay, initialFen: Option[Fen.Full], parsed: ParsedPgn)
-case class ImportReady2(
-    game: ChessGame,
-    result: Option[TagResult],
-    replay: Replay,
-    initialFen: Option[Fen.Full],
-    parsed: ParsedPgn
-)
-
 trait Explorer:
   def apply(id: GameId): Fu[Option[Game]]
 
 trait Divider:
   def apply(id: GameId, sans: => Vector[SanStr], variant: Variant, initialFen: Option[Fen.Full]): Division
-
-type ParseImport = (ImportData, Option[UserId]) => Either[ErrorStr, ImportReady]
-
-trait Importer:
-  val parseImport: ParseImport
-  def importAsGame(data: ImportData, forceId: Option[GameId] = none)(using Option[MyId]): Fu[Game]
 
 object PgnDump:
   case class WithFlags(
