+++ modules/study/src/test/newTreeTest.scala
@@ -5,7 +5,7 @@ import chess.format.pgn.{ Glyphs, ParsedPgn, San, Tags, PgnStr, PgnNodeData, Com
 import chess.format.{ Fen, Uci, UciCharPair, UciPath }
 import chess.MoveOrDrop.*
 
-import lila.core.game.{ ImportData, ImportReady }
+import lila.tree.{ ImportData, ImportReady }
 import lila.tree.Node.{ Comment, Comments, Shapes }
 
 import cats.syntax.all.*
