+++ modules/tree/src/main/ParseImport.scala
@@ -2,12 +2,32 @@ package lila.tree
 
 import chess.format.Fen
 import chess.format.pgn.{ ParsedPgn, Parser, PgnStr, Reader, Sans }
-import chess.*
+import chess.{ Game as ChessGame, * }
 import chess.variant.*
 import scala.util.chaining.*
 
 import lila.core.userId.UserId
-import lila.core.game.{ ImportData, ImportReady2, TagResult }
+import lila.core.game.{ Game, NewGame }
+
+case class TagResult(status: Status, winner: Option[Color]):
+  // duplicated from Game.finish
+  def finished = status >= Status.Mate
+
+case class ImportData(pgn: PgnStr, analyse: Option[String])
+case class ImportReady(game: NewGame, replay: Replay, initialFen: Option[Fen.Full], parsed: ParsedPgn)
+case class ImportReady2(
+    game: ChessGame,
+    result: Option[TagResult],
+    replay: Replay,
+    initialFen: Option[Fen.Full],
+    parsed: ParsedPgn
+)
+
+type ParseImport = (ImportData, Option[UserId]) => Either[ErrorStr, ImportReady]
+
+trait Importer:
+  val parseImport: ParseImport
+  def importAsGame(data: ImportData, forceId: Option[GameId] = none)(using Option[MyId]): Fu[Game]
 
 private val maxPlies = 600
